Crate pc_mqtt_rs

Process Control Hyperdrive MQTT project (Rust version)

This documentation can either be read from the source code or by using the “index.html” (recommended). The latter was generated using “cargo doc” and can be found in the “docs” folder.

The project consists of 4 parts:

Steering controller (blink, speed, lane)

Emergency controller (relay)

Track the tracks controller (track, relay)

Personal addition (track, relay)

All client implementations is found within the client module and shared code is found within the library module.

Available controllers/clients

Each client module has a struct that holds some data about its purpose and a vehicle list. They all initialize a new MQTT client and run in their own thread. Since all the communication is done through MQTT, they can be mixed and matched with their counterparts written in Python. The only limitation being that the relay client needs to be started before any other client, to avoid lost connect messages.

Steering controllers

Blink

Every second it sends a message to each vehicle setting their lights to either on or off.

Speed

Every 3 seconds it sends a message to set the speed of each vehicle. The speed values are given to the controller with a list, and is iterated.

Lane

Every 5 seconds it sends a message to change lane of each vehicle by iterating a list of values given as an argument.

Emergency controller

Both the emergency and personal addition controllers are implemented inside the relay module/client. The relay client is responsible for relaying messages from every other client to the broker. It will also handle emergency messages and personal addition (zone) messages, and if necessary overwrite any speed messages.

Tracking and personal addition controllers

It receives and stores track ID numbers for each vehicle. If a vehicle is on a track that has a speed limit (velocity 200), it is appended to “slow_vehicles” list and published for the relay to take action.

Modules

client 🔒

library 🔒

Structs

Blink

Holds the current state and a list of vehicles.

ClientWrapper

Rumqttc client wrapper, wraps the client in an Arc<Mutex<>> to allow sharing it between threads safely.

ConnectionWrapper

Lane

Struct holding the offsets and a list of vehicles.

Mqtt

Relay

The Relay struct holds a list of vehicle IDs, the emergency state, a list of vehicles inside a slow zone, and the last speed value.

Speed

Struct holding lists of velocities and vehicles.

Track

Enums

Payload

An enum that holds most of the payloads/messagess used in the project.

Topic

An enum that holds almost all the topics used in the project.

Functions

blocking_emergency_handler

Blocks thread and publishes emergency messages on the keypress of enter.

connect_vehicles

Sends Connect(true) to each vehicle.

discover_vehicles

Asks for discovered vehicle IDs and prints them to stdout.

set_ctrlc_handler

Sets up a handler to disconnect vehicles on CTRL+C.